RUMM-180 Extract the NetworkInfo impl to core
