+++ dd-sdk-android/src/androidTest/kotlin/com/datadog/android/internal/utils/utils.kt
@@ -1,7 +1,7 @@
 package com.datadog.android.internal.utils
 
+import com.datadog.android.core.internal.net.info.NetworkInfo
 import com.datadog.android.log.internal.domain.Log
-import com.datadog.android.log.internal.net.NetworkInfo
 import com.datadog.android.log.internal.user.UserInfo
 import fr.xgouchet.elmyr.Forge
 import fr.xgouchet.elmyr.ForgeryException
