+++ dd-sdk-android/src/main/kotlin/com/datadog/android/core/internal/data/upload/DataUploadHandlerThread.kt
@@ -10,7 +10,7 @@ import android.os.Handler
 import android.os.HandlerThread
 import com.datadog.android.core.internal.data.Reader
 import com.datadog.android.core.internal.net.DataUploader
-import com.datadog.android.log.internal.net.NetworkInfoProvider
+import com.datadog.android.core.internal.net.info.NetworkInfoProvider
 import com.datadog.android.log.internal.system.SystemInfoProvider
 
 internal class DataUploadHandlerThread(
