+++ dd-sdk-android/src/main/kotlin/com/datadog/android/core/internal/net/info/NetworkInfo.kt
@@ -4,7 +4,7 @@
  * Copyright 2016-2019 Datadog, Inc.
  */
 
-package com.datadog.android.log.internal.net
+package com.datadog.android.core.internal.net.info
 
 internal data class NetworkInfo(
     val connectivity: Connectivity = Connectivity.NETWORK_NOT_CONNECTED,
