+++ dd-sdk-android/src/test/kotlin/com/datadog/android/log/assertj/LogAssert.kt
@@ -6,8 +6,8 @@
 
 package com.datadog.android.log.assertj
 
+import com.datadog.android.core.internal.net.info.NetworkInfo
 import com.datadog.android.log.internal.domain.Log
-import com.datadog.android.log.internal.net.NetworkInfo
 import com.datadog.android.log.internal.user.UserInfo
 import org.assertj.core.api.AbstractObjectAssert
 import org.assertj.core.api.Assertions.assertThat
